a load of stuff to deal with marking maths expressions; and some UI bits
first: the UI now has a built-in list of example marking algorithms,
generated from the files in the `algorithms` directory.

This commit mainly adds a ton of JME functions to help with manipulating
and inspecting mathematical expressions. The "mathematical expression"
marking algorithm is now implemented (as 'jme'), and there's a very
simple malrules pattern-checking example.

I had a look at how maxima simplifies expressions, and had a go at
implementing its ordering of terms. It works quite well! That work
isn't strictly related to this marking stuff, so it should really go
in a branch of the Numbas repo.